[gaim-migrate @ 5817] Phroggie says windows doesn't have a z modifier, so I'm just going to change size_t to an int, since it really doesn't matter.
I also changed some sprintf's to snprintf's, I'm not really sure why.

Changed a printf to gaim_debug in oscar.c, I don't know how that got
in there.  And I applied a patch that adds an optional callback thing
for when a connection is closed.  Gaim doesn't use it, but the guy
that submitted the patch said it was useful to him.

--HG--
extra : convert_revision : 11ee954e83fd685c23b161baaf9d77dcdc523b9b

+++ src/protocols/jabber/jabber.c
@@ -2055,7 +2055,7 @@ static void jabber_xfer_start(struct gaim_xfer *xfer)
 static size_t jabber_xfer_read(char **buffer, struct gaim_xfer *xfer) {
 	struct jabber_xfer_data *data = xfer->data;
 	char test;
-	size_t size;
+	int size;
 
 	if(read(xfer->fd, &test, sizeof(test)) > 0) {
 		data->headers = g_string_append_c(data->headers, test);
@@ -2065,7 +2065,7 @@ static size_t jabber_xfer_read(char **buffer, struct gaim_xfer *xfer) {
 			if(data->newline) {
 				gchar *lenstr = strstr(data->headers->str, "Content-Length: ");
 				if(lenstr) {
-					sscanf(lenstr, "Content-Length: %zd", &size);
+					sscanf(lenstr, "Content-Length: %d", &size);
 					gaim_xfer_set_size(xfer, size);
 				}
 				gaim_xfer_set_read_fnc(xfer, NULL);
